+++ test/addons-napi/test_constructor/test_constructor.c
@@ -6,7 +6,7 @@ napi_ref constructor_;
 void GetValue(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
 
@@ -26,7 +26,7 @@ void GetValue(napi_env env, napi_callback_info info) {
 void SetValue(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
 
@@ -46,7 +46,7 @@ void SetValue(napi_env env, napi_callback_info info) {
 void Echo(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
